src/Multimodal/requirements.txt
# 多模态CLIP训练环境依赖

# 深度学习框架
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# CLIP模型
git+https://github.com/openai/CLIP.git

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 机器学习工具
scikit-learn>=1.0.0
scipy>=1.7.0

# 数据加载和处理
tqdm>=4.62.0
h5py>=3.6.0

# 日志和监控
tensorboard>=2.8.0
wandb>=0.12.0

# 图像处理增强
albumentations>=1.1.0
imgaug>=0.4.0

# 文本处理
nltk>=3.6.0
transformers>=4.15.0

# 配置管理
PyYAML>=6.0
omegaconf>=2.1.0

# 其他工具
jupyterlab>=3.2.0
ipywidgets>=7.6.0
requests>=2.26.0
